sysvinit: Ensure that iscsid is started
vdsm is starting the iscsid daemon when it starts, but iscsid "start"
command may silently *not* start the iscsid daemon, because vdsm is
configuring it's nodes to login manually. When iscsid daemon is not
running and vdsm is connecting to a storage server, iscsiadm starts
iscsid daemon. This works when /etc/iscsi/iscsi.conf contains:

    iscsid.startup = /etc/rc.d/init.d/iscsid force-start

This is the default configuration on a fresh install of EL 6, and is
common on developers machines. However when this option is missing,
iscsiadm fails and iscsi storage is not available. According to the
comments on the bug, the missing configuration is common in the field.

This patch starts iscsid using "force-start" command, ensuring that
iscsid daemon is started and removing the dependency on iscsi.startup
configuration. This restores the behavior from commit 0233b5e952 in Nov
2011 until commit 3c0adf791e in May 2013. The behavior was changed to
make systemd-based deployment (which has no "force-start") start iscsid,
but now that we have an independent systemd service, it can be safely
avoided.

This change does not effect the deployment for systemd or upstart.

Change-Id: Idb5c3de4f8dccb1d8344182d5016feabcf86a7df
Relates-To: http://gerrit.ovirt.org/14630
Bug-Url: https://bugzilla.redhat.com/1056948
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/23601
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/23770
Reviewed-by: Allon Mureinik <amureini@redhat.com>