caps: report malfunctioning libvirt
the fix applied in commit 6c6da596336d6cc1ff121057877ac6ab1f00f147
can be a source of confusion for users which expect
a Fedora 19 node is able to join a 3.4 cluster.

This patch adds log messages which briefly explain
the reason of the incompatibility and suggest a fix.

Bug-Url: https://bugzilla.redhat.com/1056918

Change-Id: I9fd62ecaabc30ccd2c5d8b8cc441de6a44b8e670
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/23628
Reviewed-by: Dan Kenigsberg <danken@redhat.com>