spec: do not remove vdsm logs when pkg removed
In the commits 4a07386 0b6035a we introduced the concept
of creating of metadata.log, mom.log, supervdsm.log and vdsm.log
logs during the rpm install and sets the proper
user/group. However, now when removing vdsm the logs are removed too
which is not the behaviour as before these commits. This patch will
add ghost macro in the spec to keep logs available when vdsm is removed.

Change-Id: Idc8e62d5a0a6166c9141fe9d6016a525a3331694
Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/23876
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>