Adding configure sanlock on force
If sanlock is down we cannot check sanlock process, so isconfigure return that
sanlock process is already related to the supplementary group.
Configure verb won't configure sanlock if isconfigure return True currently.
Therefore, this patch change the ConfigureOnForce to True so
'vdsm-tool configure --force' will perform the usermod action as the
user expected.

Change-Id: I0bb3125e64cf02e55738f5978fd81d605dd84d38
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1057225
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/24007
Reviewed-by: Sandro Bonazzola <sbonazzo@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/24153