Validate conf in configure only if not forcing overriding
When --force flag is sent to configure the conf might be overrided
(depends on reconfigureOnForce return value).
Therefore, checking validate conf before overriding can lead to a fail
while the current configuration is not relevant.

libvirt_configure.sh already manage to set the right configure in
libvirtd.conf and qemu.conf according to the ssl verb in vdsm.conf.
BUT, without this patch we validate the current config
state before overriding the files even though the --force flag was
stated, and this lead to fail the configure.

With that patch, setting in vdsm.conf ssl=True and then running
"vdsm-tool configure --force" will update libvirtd.conf and qemu.conf
according to vdsm.conf state.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1036883
Change-Id: Ibd9e95c06b1437074f37248e77bb086a3c3bad63
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/24132
Reviewed-by: Dan Kenigsberg <danken@redhat.com>