spec: vdsm pkg for EL distro must include pkla
vdsm on EL system should ship the .pkla file or libvirt will
keep asking password and vdsm won't be able to start correctly.
Originally the commit f824941b has added into %files session
the polkit files but on a %else conditional for %0{?rhel} and
this code contains %else for %if 0%{?fedora} >= 18 which won't be EL as expected.
Additionaly, a new commit bae85d50 added %with_systemd macro and triggered
this bug inserting new %if 0%{?rhel}/%else which contains %if 0%{?fedora} >= 18
and %else but won't work too for EL.

Change-Id: I298f21ee6a3e5f2fdf78939ea421a2152490972b
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1060524
Signed-off-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/24083
Reviewed-by: Dan Kenigsberg <danken@redhat.com>