netinfo: Fix wrong filtering of dummy devices
Dummy (and veth) devices were being filtered for use by vdsm in the
method isFakeNic, but in the method that is used to determine what
is reported as nics, isNicLike, it was not being used, and instead
isDUMMY and isVETH were being called.

Before this patch, any veth and dummy device were reported in the
nics() method of netinfo.
After this patch only those that match the explicit fake setting
will be reported.

Change-Id: I3809eb1bb3230543bd586dcf722e2603da16952a
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/24175
Reviewed-by: Dan Kenigsberg <danken@redhat.com>