sp: Remove invalid validation of stale meta data
When refreshing a storage pool, we used to validate the the metadata
before refreshing the pool. This validation fail if the storage domain
was just upgraded from old format where metadata was storaged on the
medadata lv, to the new format keeping metadata in the lv tags.

The validation is pefromed now after the refresh, ensuring that the
metadata is correct after a successful refresh.

Change-Id: I27ab243a03f23b1155867f2eeec98b1481b5b72d
Bug-Url: https://bugzilla.redhat.com/954169
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/24568
Reviewed-by: Ayal Baron <abaron@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>