vdsm: Enable config of VM serial number
Added support for passing explicit serial number
as a part of vmParams from the engine.

This change will enable users to configure various "serial number
policies" on the engine such as:
- use VM's UUID
- use Host's UUID
- provide custom value
enabling greater control over the resulting VM's serial number.

Notes:
In the first case, the engine will pass the VM's UUID also as a value of
the 'serial' param.

In case this new 'serial' param is not passed we fall back to the old
behavior when we use the host UUID.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=918138
Change-Id: I76309e833f0b00e69a187037a415642a8daf35f6
Signed-off-by: Martin Betak <mbetak@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/24274
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>