xmlrpc: Support HTTP 1.1
Supporting HTTP 1.1 allow using keep-alive on the client side,
decreasing the latency and load on the engine and increasing the number
of host that engine can support.

This patch override do_POST and report_404 methods that wrongly shutdown
the connection at the end of the do_POST request. This is the
responsibility of the server, and not the request handler.  The fixed
methods are used only on Python 2.6. Python 2.7 already include these
changes. Then the protocol_version is set to HTTP/1.1, enabling
automatic keep-alive.

A new configuration option "xmlrpc_http11" can be used to disbale this
feature and use HTTP/1.0 as used before.

We can see in the logs that only few threads are created now for service
XMLRPC requests, and most requests are handled by the same thread.

Change-Id: Ie033d5c53c81c8db99d5c26697a1727be030e0b4
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/24294
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>