Fixing ssl config validation
Currently we verify only the config of livbirtd.conf and qemu.conf
when ssl=False in vdsm.conf. If ssl=True we assumed the validation
passes. This patch add a check also to ssl=True.

Except that, the validation function is not general and checks only ssl
configuration. This patch splits to specific test_conflict functions and
allow adding more test_conflict functions in the future.

Change-Id: I3d9ada1b81d275b8cac9391c15ab903ec9552184
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/24425
Reviewed-by: Dan Kenigsberg <danken@redhat.com>