netlink: Introduce socket pool
If a lot of netlink requests happened _nl_connect would fail at
the C level, more concretely when doing bind, which would return
EADDRINUSE.

This patch addresses that issue as well as reduces the the resource
consumption of the module by reusing the netlink sockets. The reusing
is done by a semaphore protected socket pool.

Bug-Url: https://bugzilla.redhat.com/1070061
Change-Id: I657ac3d3e0c2661ce73bdef9aa807ead888a42eb
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/24603
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/25056