clientIF: prepareVolumePath payload cleanup
the prepareVolumePath code path for cdrom/floppy
images is complicated due the fact the code has to
deal with both regular and payload-created
images (vmPayload, used by cloud-init).

The code has been patched various times, and in the end the
code became more complex than is needed.

This patch refactors the existing code in order to simplify
it, without changes of behaviour.

Change-Id: I058206b7506ddbb5ec087c9ea0963a10ed57affb
Relates-To: http://gerrit.ovirt.org/22928/
Relates-To: https://bugzilla.redhat.com/1047356
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/24636
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>