vdscli: remove usage of VDSMBRIDGE
When not given an address to connect, vdscli.connect() attempts to guess
the address of a local Vdsm server. With ssl disabled, it used to fall
back to looking up the address of the management bridge.

However, Vdsm server no longer care about VDSMBRIDGE when deciding on
which IP address it should listen, so the client side should loose it,
too.

Change-Id: I2127b2d81ca675f5709f8d3f7d2600df98c58be0
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/24813
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>