vdscli: remove usage of VDSMBRIDGE
When not given an address to connect, vdscli.connect() attempts to guess
the address of a local Vdsm server. With ssl disabled, it used to fall
back to looking up the address of the management bridge.

However, Vdsm server no longer care about VDSMBRIDGE when deciding on
which IP address it should listen, so the client side should loose it,
too.

This patch is being backported in order to simplify backporting of a
follow-up patch.

Change-Id: I2127b2d81ca675f5709f8d3f7d2600df98c58be0
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Related-To: https://bugzilla.redhat.com/1111075
Reviewed-on: http://gerrit.ovirt.org/24813
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/29181
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>