vdsm-reg: eliminate the VDSMBRIDGE macro
VDSMBRIDGE is a historical macro holding the "management bridge" of
Vdsm. Nowadays, this bridge may well not exist as management network are
better kept bridge-less. The concept of "management bridge" is currently
used only within vdsm-reg, and only when it needs to register to a legacy
oVirt-Engine that expects the bridge to exist.

This patch hides the unneeded macro within vdsm-reg, to be completely
dropped when vdsm-reg is replaced by an ovirt-node-register utility.

Change-Id: Iaa1dae084ae9df44c10e0f6ae43584544dab2b8d
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/24814
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Tested-by: Douglas Schilling Landgraf <dougsland@redhat.com>