xmlrpc: Disable needless and slow dns lookup
In Python 2.6, BaseHTTPServer.BaseRequestHandler is invoking
socket.getfqdn() after each request, to log the client fully qualified
hostname. If dns is not configured properly, this causes needless delay
after each request. In Python 2.7, the address_string is not used.

This patch overrides address_string to return client address, avoiding
the dns lookup, when running in Python 2.6.

Change-Id: I6b62d4a5b6765d3539c92b72e821b9029b043941
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/24871
Reviewed-by: Dan Kenigsberg <danken@redhat.com>