Install vdsm.conf.example during build
The sample conf is installed by spec under vdsm docs. Docs files are not part
of RHEV-H installation to keep rhev-h env small as possible.
Although, the sample file was delivered only on Fedora which is the
standard for package delivering (to publish sample conf with the
package installation).
The sample file was requested also in rhev-h by users for reference.
This patch moves the file installation from spec to create it during
make.

Change-Id: I2606c571cd6de06eee811bbd340a790f98d11907
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1023941
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/24945
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>