ifcfg: Do not take bond/nic down when removing one of their multiple nets
When taking removing one of the multiple networks that sit on top of
a bond/nic, it is necessary to adjust the mtu of the devices to the
new maximum required by the remaining networks that use the devices.
Example
    netA has mtu 2500 and is set over bond5.10
    netB has mtu 2000 and is set over bond5.10

    when both are set, bond5.10, bond5 and its link devices will
    have mtu 2500. When netA is removed that should be adjusted
    to 2000.

For ifcfg to apply the adjustment, it was necessary to ifdown
the nic/bond and up it again so it would read the new configuration.
The problem is that the netB of the example would see a loss of
connectivity which could affect the users. This patch addresses that
by using the ipwrapper to make the mtu adjustment and only take
down the device when no remaining networks use it.

Change-Id: Ifffad45193ccd047acdefed7412bc9ce6303874a
Bug-Url: https://bugzilla.redhat.com/1060781
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/25003
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/25680