ifcfg.removeNic: ignore missing host interface
There may be cases where the Ifcfg configurator is asked to remove a Nic
that is no longer exist on the host. This can happen if the device was
renamed by an administrator or if it was unplugged from the system.

In this patch we do not try to ifup the missing device.

Change-Id: Ia892202b837f4515cd44a3b485fbe1b2a447457c
Bug-Url: https://bugzilla.redhat.com/1059773
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/25009
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>