caps: Collect numa information
This patch modifies the caps module to add function of collect
host numa information and report to ovirt engine.
It will add three keys in the rpc response of getCapabilities
whose value are the host's numa related information.

Key numaNodes, the value is a map, each item follows this format:
{'nodeIndex':{'cpus':['int'], 'totalMemory':int}}
nodeIndex      - The index of numa node
cpus           - The cpu ids(see the cpu id in libvirt api's
                 capabilities function) which belong to this
                 numa node
totalMemory    - The total memory of this numa node in MB

Key numaNodeDistance, the value is a map, each item follows this
format:
{'nodeIndex':[distanceList]}
nodeIndex      - The index of numa node
distanceList   - Distances from self to other nodes in sequence,
                 including self to self

Key autoNumaBalancing, the value is a int, represents the status
of auto numa balancing function.

Change-Id: I63eeb697ab986c3b9cad0dc44f41924f329e52cd
Bug-Url: https://bugzilla.redhat.com/1069303
Signed-off-by: Bruce Shi <xiao-lei.shi@hp.com>
Reviewed-on: http://gerrit.ovirt.org/23703
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Martin Sivák <msivak@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>