recovery: try to restore VMs from recovery files
on recovery, VDSM tries to recover all the VMs reported
as active by libvirt.
If a VM disappears while VDSM is down or restarting, the
VM silently disappears without VDSM noticing, and the recovery
file is discarded.

This patch adds a second recovery try for the VMs which have
a recovery file but not yet restored using the standard recovry
loop.

Those VMs are most down on the recovering host, but VDSM has
to notify the engine so it is can do proper cleanup and it can
ensure the correct state is reported to users.

This patch depends on
I00ef12883c8035209de0f273925eb8603d6b6da8
which make sure VMs are setup and reported correctly if
the underlying QEMU process disappeared.

Change-Id: Id495f6047ba658c2b04da19bd7bf76425b3b9659
Bug-Url: https://bugzilla.redhat.com/1045626
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/25276
Reviewed-by: Dan Kenigsberg <danken@redhat.com>