setupNetworks: Fix incorrect mtu overriding when adding multiple networks
When configuring several networks over a bond, if the iteration over
the networks to be added made networks with lower MTUs to be added
after those with higher ones, the last mtu, regardless of value would
be set.

This was because the _netinfo object is passed from addNetworks to
addNetworks and the objectivize of the succeeding addNetworks would
not see the higher MTUs set by the preceding addNetworks. This would
only affect the bond MTU when some bond setting would be detected
as changed and it would trigger bond reconfiguring (for a list of
what is checked for changes, see netmodels.py:Bond.configure).

Change-Id: Ia375c2e0e4a1896ab99d734e3203a5ef49570f36
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1072411
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/25343
Reviewed-by: Dan Kenigsberg <danken@redhat.com>