interface xml: allow Engine to specify number of queues
If a vNIC is used for in-host communication between VMs, its throughput
would gain if it is allocated with multiple queues, numbering as the
vCPUs of the VM. This comes with a price of pinning host memory for this
purpose. Cf. http://www.linux-kvm.org/page/Multiqueue.

This patch exposes the "queues" per-device custom property which can be
set in a vNIC profile in Engine. Enable it there with

    engine-config -s "CustomDeviceProperties={type=interface;prop={
        other-nic-properties;queues=[0-9]*}}"

Note that this property is currently supported only by libvirt >= 1.0.6.
Older versions silently ignores it.

Change-Id: I83e129e3c6900190f9b5d863d832745833fe2699
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/25390
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>