sp: fix spm start when failing to produce domain
During spmStart we attept to go over all the domains and set their role
to be regular in case it's not to avoid situation in which there are two
domains with the master role.
If we fail to produce any of the domains (which might happen, the
domains returned by getDomains(activeOnly=True) we shouldn't fail,
as we attempt to produce all the domains that are marked as active in the
metadata, but it doesn't mean that they are actually reachable and that
calling produce() for the domain won't fail.

In case the we fail to call produce() on domain with
StorageDomainDoesNotExist exception, it doesn't mean that spmStart should
fail - otherwise we might never succeed to start spm.

I believe that we should ignore any exception to avoid the same case in
different scenarios (for example - failure to alter the metadata), but
as reviewers requested, let's start with StorageDomainDoesNotExist
exception only.

Change-Id: Ia4f34360770ca8c8741a50956d0cab92bcd9a810
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1072900
Signed-off-by: Liron Aravot <laravot@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/27194
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>