Avoid to recompile namedtuple ATTR classes in lvm.
Re-defining the ??_ATTR classes may be hurting the performance
during the VM creation (preliminary). This patch avoids attribute
class creation each time the lvm object is recreated.

Should be noted that re-creating lvm cached objects during the VM
creation is not required, therefore paths modified by the actual
patch should not be called in such scenario.
This abnormal situation will be corrected in a following patch.

Dealing with variable attribute strings is not required for the
actual development of oVirt and the lvm module, then
_attr2NamedTuple() can be removed.

Change-Id: I66110a7f25fb5cfd80ddffe9a22e1cbac11de447
Signed-off-by: Eduardo <ewarszaw@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/25678
Tested-by: Meital bourvine <meitalbourvine@gmail.com>
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Yeela Kaplan <ykaplan@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Saggi Mizrahi <smizrahi@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>