net_persistence: Make unified persistence default
Up until now, we've been using the legacy ifcfg persistence to store
and retrieve network configurations. This patch switches the default
to 'unified' so that we'll be able to depend on it moving forward.

Unified persistence stores its persistence in three places:
- In transaction data: memory,
- Current configuration: /var/run/vdsm/netconf/{nets,bonds}
- Persisted configuration:
      /var/lib/vdsm/persistence/netconf/{nets,bonds}

Configurators are responsible to implement a flush method that cleans
the system from its own persistence files.

Change-Id: I836972041826e4af47e83ef5097994a5203fd859
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/25784
Reviewed-by: Livnat Peer <lpeer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>