LiveMerge: Add merge verb and related machinery
This patch enables the submission of an actual live merge operation to
the underlying VM.  Once submitted, we poke the stats thread to make
sure block job information is updated and persist the given merge
parameters in the VM Disk conf.  This ensures that in case vdsm restarts
during a block job we will not forget to handle it the next time we are
started.

Limitations:
 - Does not yet work when the active layer is involved due to [1]

[1] https://bugzilla.redhat.com/show_bug.cgi?id=1062142

Change-Id: I077e1bdf0130ebdc218135ca717a6a810bbab12c
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/26635
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>