virt: migration: always run migration monitor
Before this patch the migration monitor thread is always
run, unless the migration monitor interval configuration option
is explicitely set to zero.

This patch moves this check inside the migration.Monitor class.
This patch provides admittedly little benefit on its own,
in the form of minor cleanup, but is a step toward the unification
of monitor and downtime threads.

Change-Id: Ib4953ddda5a5c6c0ecd7ea0f95377309e18f771a
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/25975
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>