storage: Add logging on filesystem altering operations
The rationale behind this patch is to add logging on
filesystem altering operations. Before, no trace was given when creating or
deleting symbolic links or removing or renaming files and folders.
This patch tries to solve this adding debug messages before the
operations are executed.

This patch only includes the logging in the VDSM process so no out of
process log management is needed.

Change-Id: I3602513af123951f71091c03f799e36ea759aa61
Signed-off-by: Xavi Francisco <xfrancis@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/26046
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>